# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This api will help you control your links from outside the site",
        "title": "Short URL Generator API",
        "contact": {
            "name": "API Support",
            "url": "https://nktkln.com",
            "email": "nktkln@nktkln.com"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/add": {
            "post": {
                "description": "Reference abbreviation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/others.ApiUrl"
                        }
                    }
                }
            }
        },
        "/archive": {
            "get": {
                "description": "Reference abbreviation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Archive",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/others.Api"
                        }
                    }
                }
            }
        },
        "/delete": {
            "post": {
                "description": "Delete url from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/others.Api"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "others.Api": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                }
            }
        },
        "others.ApiUrl": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}